Modify the swap and backup directories
Swap files in the working directory can annoying when vim exits
abnormally as it can be difficult and time consuming to detect where the
swap files might have been left - they are only detected when trying to
open the same file later on and usually it is not the right time for
that. It also gets in the way of opening the same file in two vim
sessions.

Also, relativenumber is now used in addition to number - because it
looks cool.