Restore the original swap default of current dir
The ~/.vim/swap directory is still kept as a backup, but the current
directory is now the default as it used to be.

Alse, the sensible settings of polyglot were disabled as they disabled
the swap warnings. To do this, the setting had to be set before loading
the plugin, so an early-settings directory was created which is sourced
before the plugins.